Pc.activation

+++ ruleskit/condition/hyperrectanglecondition.py
@@ -1,16 +1,15 @@
 from typing import List, Union
 import numpy as np
 from .condition import Condition
-from rule.activation import Activation
+from ..activation.activation import Activation
 
 
 class HyperrectangleCondition(Condition):
-
     def __init__(
         self,
         features_indexes: Union[List[int], None] = None,
-        bmins: Union[List[float], None] = None,
-        bmaxs: Union[List[float], None] = None,
+        bmins: Union[List[int], None] = None,
+        bmaxs: Union[List[int], None] = None,
         features_names: Union[List[str], None] = None,
         empty: bool = False,
     ):
@@ -31,9 +30,11 @@ class HyperrectangleCondition(Condition):
                 self._features_names = ["X_" + str(i) for i in self._features_indexes]
 
     def __and__(self, other: "HyperrectangleCondition"):
-        args = [i+j for i, j in zip(self.getattr, other.getattr)]
-        return HyperrectangleCondition(features_indexes=args[0], bmins=args[1], bmaxs=args[2],
-                                       features_names=args[3], empty=False)
+        args = [i + j for i, j in zip(self.getattr, other.getattr)]
+        # noinspection PyTypeChecker
+        return HyperrectangleCondition(
+            features_indexes=args[0], bmins=args[1], bmaxs=args[2], features_names=args[3], empty=False
+        )
 
     def __add__(self, other: "HyperrectangleCondition"):
         return self & other
